+++ src/main/java/teamroots/embers/mixin/embersconstruct/OreDictValuesMixin.java
@@ -2,6 +2,7 @@ package teamroots.embers.mixin.embersconstruct;
 
 import com.peatral.embersconstruct.util.OreDictValues;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import teamroots.embers.config.ConfigMachine;
@@ -25,39 +26,33 @@ public class OreDictValuesMixin {
         return ConfigMachine.STAMPER.stampGearAmount;
     }
 
-    // Redirect all returns to actual config values
-    @Redirect(method = "lambda$static$0", at = @At(value = "INVOKE", target = "Ljava/lang/Integer;valueOf(I)Ljava/lang/Integer;"))
-    private static Integer amount_ingot(int i) {
-        return ConfigMisc.ingotFluidAmount;
-    }
-
-    @Redirect(method = "lambda$static$1", at = @At(value = "INVOKE", target = "Ljava/lang/Integer;valueOf(I)Ljava/lang/Integer;"))
-    private static Integer amount_ore(int i) {
-        return ConfigMisc.ingotFluidAmount * ConfigMachine.MELTER.melterOreAmount;
-    }
-
-    @Redirect(method = "lambda$static$2", at = @At(value = "INVOKE", target = "Ljava/lang/Integer;valueOf(I)Ljava/lang/Integer;"))
-    private static Integer amount_block(int i) {
-        return ConfigMisc.blockFluidAmount;
-    }
-
-    @Redirect(method = "lambda$static$3", at = @At(value = "INVOKE", target = "Ljava/lang/Integer;valueOf(I)Ljava/lang/Integer;"))
-    private static Integer amount_nugget(int i) {
-        return ConfigMisc.nuggetFluidAmount;
-    }
-
-    @Redirect(method = "lambda$static$4", at = @At(value = "INVOKE", target = "Ljava/lang/Integer;valueOf(I)Ljava/lang/Integer;"))
-    private static Integer amount_dust(int i) {
-        return ConfigMisc.ingotFluidAmount;
-    }
-
-    @Redirect(method = "lambda$static$6", at = @At(value = "INVOKE", target = "Ljava/lang/Integer;valueOf(I)Ljava/lang/Integer;"))
-    private static Integer amount_plate(int i) {
-        return ConfigMisc.ingotFluidAmount * ConfigMachine.STAMPER.stampPlateAmount;
-    }
-
-    @Redirect(method = "lambda$static$8", at = @At(value = "INVOKE", target = "Ljava/lang/Integer;valueOf(I)Ljava/lang/Integer;"))
-    private static Integer amount_gear(int i) {
-        return ConfigMisc.ingotFluidAmount * ConfigMachine.STAMPER.stampGearAmount;
+    /**
+     * @author MasterEnderman
+     * @reason Use actual config values
+     */
+    @Overwrite
+    public int getValue() {
+        OreDictValues odv = (OreDictValues) (Object) this;
+        switch (odv) {
+            case INGOT:
+            case DUST:
+                return ConfigMisc.ingotFluidAmount;
+            case ORE:
+                return ConfigMisc.ingotFluidAmount * ConfigMachine.MELTER.melterOreAmount;
+            case BLOCK:
+                return ConfigMisc.blockFluidAmount;
+            case NUGGET:
+                return ConfigMisc.nuggetFluidAmount;
+            case SHARD:
+                return 72;
+            case PLATE:
+                return ConfigMisc.ingotFluidAmount * ConfigMachine.STAMPER.stampPlateAmount;
+            case GEM:
+                return 666;
+            case GEAR:
+                return ConfigMisc.ingotFluidAmount * ConfigMachine.STAMPER.stampGearAmount;
+            default:
+                return 0;
+        }
     }
 }